Attempt to fix rsync not working on jenkins
For some reason a gusteau job running on jenkins is bailing out 10% of
the way through an initial rsync.

It doesn't happen when the job is run on the command line - only through
jenkins.

Will try making the rsync output less verbose here by removing the "P"
option to see if that fixes it.