Just use `tee -a` to log the LOGFILE
I was hoping this `chef-solo -L` would still log to STDOUT aswell as the logfile,
but unfortunately it only logs to the logfile.

Am I using the wrong version of chef-solo (ie. v11.4.4),
or is it just that the documentation on this is a little confusing on
this point:

--- snip http://docs.opscode.com/ctl_chef_solo.html ---
The location in which log file output files will be saved.
If this location is set to something other than STDOUT,
standard output logging will still be performed
(otherwise there would be no output other than to a file).
--- snip http://docs.opscode.com/ctl_chef_solo.html ---